# openapi.yaml
openapi: 3.0.0
info:
  title: Vehicle Booking API
  description: API for Vehicle booking and managing Rides.
  version: "1.0"
paths:
  /appointments:
    get:
      summary: List all Booking or filter by upcoming and/or email
      parameters:
        - in: query
          name: upcoming
          schema:
            type: string
          description: Filter by upcoming Bookings
        - in: query
          name: email
          schema:
            type: string
          description: Filter by email address
      responses:
        '200':
          description: A list of Booking
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
    post:
      summary: Create a new Booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Appointment'
      responses:
        '201':
          description: The Booking was successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
  /appointments/{id}:
    get:
      summary: Get a single Booking by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: Unique identifier of the Booking
      responses:
        '200':
          description: Details of an Booking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
    put:
      summary: Update an Booking
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: Unique identifier of the Booking to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Appointment'
      responses:
        '200':
          description: The Ride was successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
    delete:
      summary: Delete an Booking
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: Unique identifier of the Booking to delete
      responses:
        '200':
          description: The Booking was successfully deleted
components:
  schemas:
    Appointment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        service:
          type: string
        phoneNumber:
          type: string
          format: phone
        appointmentDate:
          type: string
          format: date-time
        email:
          type: string
          format: email
      required:
        - name
        - service
        - phoneNumber
        - email
        - appointmentDate